refactor: 적용된 이벤트 혜택 정보의 세부 정보 패키징 변경에 따라 이름 수정
EventResult.java
DiscountAmountDto -> DiscountAmount.java
DiscountEventResultDto -> DiscountEventResult.java
GiveawayEventResultDto -> GiveawayEventResult.java
