+++ src/main/java/christmas/domain/model/classes/eventResult/DiscountAmount.java
@@ -1,0 +1,4 @@
+package christmas.domain.model.classes.eventResult;
+
+public record DiscountAmount(int discountAmount) {
+}
