+++ src/main/java/christmas/domain/model/classes/eventResult/impl/GiveawayEventResult.java
@@ -1,14 +1,14 @@
-package christmas.dto.eventResult.impl;
+package christmas.domain.model.classes.eventResult.impl;
 
 import christmas.domain.model.enums.Giveaway;
-import christmas.dto.eventResult.EventResult;
+import christmas.domain.model.classes.eventResult.EventResult;
 
 import java.util.Map;
 
-public class GiveawayEventResultDto implements EventResult<Map<Giveaway, Integer>> {
+public class GiveawayEventResult implements EventResult<Map<Giveaway, Integer>> {
     private final Map<Giveaway, Integer> giveaway;
 
-    public GiveawayEventResultDto(Map<Giveaway, Integer> giveaway) {
+    public GiveawayEventResult(Map<Giveaway, Integer> giveaway) {
         this.giveaway = Map.copyOf(giveaway);
     }
