+++ src/main/java/christmas/domain/policy/SpecialDayDiscountPolicy.java
@@ -2,9 +2,9 @@ package christmas.domain.policy;
 
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
 import christmas.domain.model.enums.AppliedEventPolicyName;
-import christmas.dto.eventResult.DiscountAmountDto;
-import christmas.dto.eventResult.EventResult;
-import christmas.dto.eventResult.impl.DiscountEventResultDto;
+import christmas.domain.model.classes.eventResult.DiscountAmount;
+import christmas.domain.model.classes.eventResult.EventResult;
+import christmas.domain.model.classes.eventResult.impl.DiscountEventResult;
 
 import java.time.LocalDate;
 import java.util.Map;
@@ -38,6 +38,6 @@ class SpecialDayDiscountPolicy {
     }
 
     private static EventResult calculateDiscountEventResult() {
-        return new DiscountEventResultDto(new DiscountAmountDto(SPECIAL_DISCOUNT_AMOUNT));
+        return new DiscountEventResult(new DiscountAmount(SPECIAL_DISCOUNT_AMOUNT));
     }
 }
