--- src/main/java/christmas/dto/eventResult/DiscountAmountDto.java
@@ -1,4 +1,0 @@
-package christmas.dto.eventResult;
-
-public record DiscountAmountDto(int discountAmount) {
-}
