--- src/main/java/christmas/dto/eventResult/impl/DiscountEventResultDto.java
@@ -1,17 +1,0 @@
-package christmas.dto.eventResult.impl;
-
-import christmas.dto.eventResult.DiscountAmountDto;
-import christmas.dto.eventResult.EventResult;
-
-public class DiscountEventResultDto implements EventResult<DiscountAmountDto> {
-    private final DiscountAmountDto discountAmount;
-
-    public DiscountEventResultDto(DiscountAmountDto discountAmount) {
-        this.discountAmount = discountAmount;
-    }
-
-    @Override
-    public DiscountAmountDto getResult() {
-        return discountAmount;
-    }
-}
